Time: 1819 ms (8.11%), Space: 52.1 MB (79.04%) - LeetHub

+++ 3573-best-time-to-buy-and-sell-stock-v/3573-best-time-to-buy-and-sell-stock-v.cpp
@@ -22,7 +22,7 @@ class Solution {
         }
 
         long long  ans1 ,ans2,ans3;
-        ans1=ans2=ans3=LONG_MIN;
+        ans1=ans2=ans3=INT_MIN;
 
         // skip this day
